requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (sqlite3 is part of Python standard library)

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and data handling
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling and validation
aiofiles==23.2.0
python-magic==0.4.27

# HTTP client for external requests
httpx==0.25.2

# Jinja2 for templating
jinja2==3.1.2

# Logging and monitoring
structlog==23.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0